requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
# Needed at runtime by AutoGluon (`pkg_resources`).
setuptools>=65.0.0
pymongo==4.6.0
motor==3.3.2
minio==7.2.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.4
python-dotenv>=1.0.0
pydantic-settings==2.1.0
passlib==1.7.4
python-jose==3.3.0
bcrypt>=3.1.7
aiofiles==23.2.1
aiokafka==0.8.0
kafka-python==2.0.2
email-validator>=1.0.5
aiohttp==3.9.1
openpyxl==3.1.2
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
imbalanced-learn==0.11.0
river==0.21.0
requests>=2.22.0
# For concept drift consumer (AutoGluon tabular retraining; backends needed to load saved predictors)
autogluon.tabular>=1.0.0,<1.2
catboost>=1.0.0
lightgbm>=4.0.0
xgboost>=2.0.0
torch>=2.0.0
fastai>=2.7.0